# openapi.yaml
...
openapi: 3.0.3
info:
  title: Formulary
  version: 0.1.0
  description: >-
    Use-case formulary for insurance AI investment. Admits candidates only with a
    complete functionality, data, and need formula, positions each on the value map,
    gates them on data readiness and conduct review before build, fixes kill
    criteria at admission, and closes each use case against benefit confirmed by
    finance in loss-ratio and expense-ratio terms.
  x-product-slug: ai-ins-formulary
  x-source: ai-in-insurance/deloitte-artificial-intelligence-in-insurance/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/conduct-gate-reviews/{reviewId}/decision:
    parameters:
      - $ref: "#/components/parameters/ReviewId"
    post:
      operationId: decideConductGate
      tags: [Gates]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [officerId, outcome, rationale]
              properties:
                officerId:
                  type: string
                outcome:
                  type: string
                  enum: [approved, approved_with_conditions, withheld]
                rationale:
                  type: string
                conditions:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Gate decision recorded and binding on build
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeConductGateReview"
...
  /v1/registry/models:
    get:
      operationId: listModelRegistrations
      tags: [Registry]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: modelFamily
          in: query
          schema:
            $ref: "#/components/schemas/ModelFamily"
        - name: unowned
          in: query
          schema:
            type: boolean
            description: Filter to production models with no registered owner
      responses:
        "200":
          description: Registered production models with owners and dependencies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeModelRegistration"
    post:
      operationId: registerModel
      tags: [Registry]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelRegistrationCreate"
      responses:
        "201":
          description: Model registered against its use cases and gate outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeModelRegistration"
...
  /v1/portfolio/analogues:
    get:
      operationId: listAnalogueCases
      tags: [Portfolio]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: sourceIndustry
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cross-industry cases with stated transfer assumptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeAnalogueCase"
    post:
      operationId: recordAnalogueCase
      tags: [Portfolio]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalogueCaseCreate"
      responses:
        "201":
          description: Analogue recorded with its explicit transfer assumption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeAnalogueCase"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ModelFamily:
      type: string
      enum:
        - speech_voice_recognition
        - sentiment_detection
        - recommendation_engine
        - text_analytics_nlp
        - pattern_anomaly_detection
        - automatic_decision_management
        - natural_language_generation
        - object_detection
        - biometrics
    Touchpoint:
      type: string
      description: Declared regulated touchpoints that trigger the conduct gate
      enum:
        - rating_factor
        - renewal_price
        - underwriting_acceptance
        - claims_decision
        - fraud_referral
        - reserving_input
        - marketing_targeting
        - none
...
    ConductGateReview:
      type: object
      required: [id, candidateId, triggeredBy, outcome]
      properties:
        id:
          type: string
        candidateId:
          type: string
        triggeredBy:
          type: array
          items:
            $ref: "#/components/schemas/Touchpoint"
        proposedFeatures:
          type: array
          items:
            type: string
        outcome:
          type: string
          enum: [open, approved, approved_with_conditions, withheld]
        officerId:
          type: string
        actuarialReviewerId:
          type: string
        conditions:
          type: array
          items:
            type: string
        rationale:
          type: string
        findings:
          type: array
          items:
            $ref: "#/components/schemas/FeatureRiskFinding"
        decidedAt:
          type: string
          format: date-time
...
    FeatureRiskFinding:
      type: object
      required: [feature, riskType]
      properties:
        feature:
          type: string
          example: residential_postcode_cluster
        riskType:
          type: string
          enum:
            - proxy_for_protected_characteristic
            - price_sensitivity_signal
            - unexplainable_to_policyholder
            - provenance_unverifiable
            - not_permitted_in_rating
        severity:
          type: string
          enum: [low, medium, high, blocking]
        mitigation:
          type: string
        raisedBy:
          type: string
...
    ModelRegistration:
      type: object
      required: [id, name, modelFamily, ownerId]
      properties:
        id:
          type: string
        name:
          type: string
        modelFamily:
          $ref: "#/components/schemas/ModelFamily"
        ownerId:
          type: string
        candidateIds:
          type: array
          items:
            type: string
        dataDependencyIds:
          type: array
          items:
            type: string
        conductGateReviewId:
          type: string
        touchpoints:
          type: array
          items:
            $ref: "#/components/schemas/Touchpoint"
        inOwnRiskAndSolvencyScope:
          type: boolean
        registeredAt:
          type: string
          format: date-time
    ModelRegistrationCreate:
      type: object
      required: [name, modelFamily, ownerId]
      properties:
        name:
          type: string
        modelFamily:
          $ref: "#/components/schemas/ModelFamily"
        ownerId:
          type: string
        candidateIds:
          type: array
          items:
            type: string
        dataDependencyIds:
          type: array
          items:
            type: string
        conductGateReviewId:
          type: string
        touchpoints:
          type: array
          items:
            $ref: "#/components/schemas/Touchpoint"
...
    AnalogueCase:
      type: object
      required: [id, sourceOrganisation, sourceIndustry, transferAssumption]
      properties:
        id:
          type: string
        sourceOrganisation:
          type: string
        sourceIndustry:
          type: string
        reportedResult:
          type: string
          example: maintenance lead times shortened by 90 percent
        underlyingFormula:
          type: string
        transferAssumption:
          type: string
          description: What must be true for this result to transfer to insurance
        insuranceTranslation:
          type: string
        recordedBy:
          type: string
    AnalogueCaseCreate:
      type: object
      required: [sourceOrganisation, sourceIndustry, transferAssumption]
      properties:
        sourceOrganisation:
          type: string
        sourceIndustry:
          type: string
        reportedResult:
          type: string
        underlyingFormula:
          type: string
        transferAssumption:
          type: string
        insuranceTranslation:
          type: string
...
    DataEnvelopeConductGateReview:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ConductGateReview"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeModelRegistration:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ModelRegistration"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeAnalogueCase:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AnalogueCase"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeModelRegistration:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ModelRegistration"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeAnalogueCase:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AnalogueCase"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"