# platform/api-server/src/domains/gates/handlers/decision.handlers.ts
/**
 * Handlers for Decision
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: gates
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { GatesDomainModule } from "../dependencies/gates-ddd.dependencies.js";
import type { DecideConductGateInput } from "@formulary/services/gates";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/conduct-gate-reviews/{reviewId}/decision
 * Operation: decideConductGate
 */
export async function decideConductGate(
  request: FastifyRequest<{ Params: { reviewId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: GatesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    reviewId: request.params.reviewId,
    ...(request.body ?? {}),
  } as DecideConductGateInput;
  const result = await deps.useCases.decisions.get.execute(input);
  return reply.code(200).send(result);
}