# platform/api-server/src/domains/portfolio/handlers/analogue.handlers.ts
/**
 * Handlers for Analogue
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: portfolio
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { PortfolioDomainModule } from "../dependencies/portfolio-ddd.dependencies.js";
import type { ListAnalogueCasesInput, RecordAnalogueCaseInput } from "@formulary/services/portfolio";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/portfolio/analogues
 * Operation: listAnalogueCases
 */
export async function listAnalogueCases(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], sourceIndustry?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: PortfolioDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    ...(request.query ?? {}),
  } as ListAnalogueCasesInput;
  const result = await deps.useCases.analogues.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/portfolio/analogues
 * Operation: recordAnalogueCase
 */
export async function recordAnalogueCase(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: PortfolioDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    ...(request.body ?? {}),
  } as RecordAnalogueCaseInput;
  const result = await deps.useCases.analogues.get.execute(input);
  return reply.code(201).send(result);
}