# platform/api-server/src/domains/registry/handlers/registry-model.handlers.ts
/**
 * Handlers for RegistryModel
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: registry
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { RegistryDomainModule } from "../dependencies/registry-ddd.dependencies.js";
import type { ListModelRegistrationsInput, RegisterModelInput } from "@formulary/services/registry";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/registry/models
 * Operation: listModelRegistrations
 */
export async function listModelRegistrations(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], modelFamily?: string | string[], unowned?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: RegistryDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    ...(request.query ?? {}),
  } as ListModelRegistrationsInput;
  const result = await deps.useCases.models.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/registry/models
 * Operation: registerModel
 */
export async function registerModel(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: RegistryDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    ...(request.body ?? {}),
  } as RegisterModelInput;
  const result = await deps.useCases.models.create.execute(input);
  return reply.code(201).send(result);
}